fixes #25: Format xmb should create xtb files for translations
