+++ package.json
@@ -12,7 +12,7 @@
     "node": ">=6.9"
   },
   "scripts": {
-    "build": "tsc -p ./src/tsconfig.json && cpx ./src/xliffmerge/xliffmerge ./dist/xliffmerge",
+    "build": "tsc -p ./tsconfig.json && cpx ./src/xliffmerge/xliffmerge ./dist/xliffmerge",
     "xliffmerge": "npm link && xliffmerge",
     "pretest": "npm run build",
     "test": "./node_modules/.bin/jasmine-node dist",
@@ -49,6 +49,6 @@
   "dependencies": {
     "chalk": "^1.1.3",
     "commander": "^2.9.0",
-    "ngx-i18nsupport-lib": "^0.1.3"
+    "ngx-i18nsupport-lib": "^1.2.0"
   }
 }
