+++ test/testdata/i18n/ngExtractedMaster1.xlf
@@ -6,11 +6,23 @@
         <source>Nachrichten</source>
         <target/>
         <note priority="1" from="description">data.edit.newstitle</note>
+        <context-group purpose="location">
+          <context context-type="sourcefile">/src/app/test-data.component.ts</context>
+          <context context-type="linenumber">19</context>
+        </context-group>
+        <context-group purpose="location">
+          <context context-type="sourcefile">/src/app/test-data.component.ts</context>
+          <context context-type="linenumber">20</context>
+        </context-group>
       </trans-unit>
       <trans-unit id="78eab955529ba0f1817c84991d9175f55bfdf937" datatype="html">
         <source>Bearbeiten Beschreibung</source>
         <target/>
         <note priority="1" from="description">data.edit.titles.description</note>
+        <context-group purpose="location">
+          <context context-type="sourcefile">/src/app/test-data.component.ts</context>
+          <context context-type="linenumber">20</context>
+        </context-group>
       </trans-unit>
       <trans-unit id="a52ba049c16778bdb2e5a19a41acaadf87b104dc" datatype="html">
         <source>Beschreibung zu <x id="INTERPOLATION"/> (<x id="INTERPOLATION_1"/>)</source>
